requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

apicapi
oslo.concurrency>=1.8.0,<1.9.0

-e git+https://github.com/noironetworks/python-opflex-agent.git@stable/juno#egg=python-opflexagent-agent
